+++ src/options-ui.tsx
@@ -1,3 +1,5 @@
+import { useEffect, useReducer } from 'react';
+import { onLanguageChange } from './languages';
 import { UrlApp } from './ui/url';
 import { LangApp } from './ui/language';
 import { TitleApp } from './ui/title';
@@ -8,10 +10,15 @@ import { LuaApp } from './ui/lua';
 import { KeymapApp } from './ui/keymap';
 
 function OptionsUI() {
+  const [generation, bump] = useReducer((value: number) => value + 1, 0);
+
+  useEffect(() => {
+    onLanguageChange(bump);
+  }, []);
 
   return (
     <>
-      <div className='h-dvh flex flex-col'>
+      <div className='h-dvh flex flex-col' key={generation}>
         <div className='prose flex flex-col p-6 gap-6 mx-auto grow'>
           <ThemeApp />
           <TitleApp />