+++ src/options.tsx
@@ -1,10 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import OptionsUI from './options-ui.tsx'
+import { initI18n } from './languages.ts'
 import './css/options.css'
 
-ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <OptionsUI />
-  </React.StrictMode>,
-)
+initI18n().then(() => {
+  ReactDOM.createRoot(document.getElementById('root')!).render(
+    <React.StrictMode>
+      <OptionsUI />
+    </React.StrictMode>,
+  )
+})